Update functions-create-first-azure-function-azure-cli.md
Added direction for Anaconda python users and heterogeneous environment users. 

Most Python users are, at this point, using the Anaconda distribution, which shouldn't be used from the plain command prompt: function deployment will fail confusingly on an SSL error. See e.g. https://stackoverflow.com/questions/45954528/pip-is-configured-with-locations-that-require-tls-ssl-however-the-ssl-module-in

Will submit a related PR to the VSCode instructions cautioning on the same topic.